+++ src/components/CalorieSpan.tsx
@@ -1,3 +1,3 @@
-import { toIntString } from "../model/calorieFunction";
+import { toIntString } from "./toIntString";
 
 export const CalorieSpan = (props: { value: number; }) => <span>{toIntString(props.value)}{' '}Cal.</span>;
