+++ src/components/input-form/FoodInputForm.tsx
@@ -2,7 +2,8 @@ import Button from "react-bootstrap/Button";
 import Col from "react-bootstrap/Col";
 import Form from "react-bootstrap/Form";
 import Row from "react-bootstrap/Row";
-import { calcFoodCalories, toIntString } from "../../model/calorieFunction";
+import { calcFoodCalories } from "../../model/calorieFunction";
+import { toIntString } from "../toIntString";
 import { Food } from "../../model/Food";
 import { VariantDanger, VariantPrimary, VariantSecondary } from "../ButtonVariant";
 import { FoodDescriptionComboBox } from "./FoodDescriptionComboBox";
