+++ src/components/toIntString.ts
@@ -1,0 +1,6 @@
+import numeral from "numeral";
+
+
+export function toIntString(val?: number): string {
+  return numeral(val).format('0');
+}
